Add OpenAI [gpt-4o-mini] support
