+++ src/InsightBoard/chatbot/__init__.py
@@ -4,3 +4,6 @@ from InsightBoard.chatbot.layout import (  # noqa: F401
     set_table,
     is_chatbot_enabled,
 )
+from InsightBoard.chatbot.dc_base import (  # noqa: F401
+    DataChat_Providers,
+)
