+++ src/InsightBoard/chatbot/prompts.py
@@ -13,8 +13,7 @@ def sql_template(tables):
     """
 
     s_preamble = """
-""You are a research assistant in charge of writing SQL queries for data stored in an SQLite database. You must ONLY respond with valid SQL queries. Explanations, if necessary, should be in the form of an SQL table containing one column labeled 'response'.
-"""
+""You are a research assistant in charge of writing SQL queries for data stored in an SQLite database. You must ONLY respond with valid SQL queries."""
     s_table = """
 The structure of the table '{table}' is defined by the following JSON schema:
 {schema}
@@ -23,9 +22,11 @@ The structure of the table '{table}' is defined by the following JSON schema:
 Key rules you must follow:
 - Do not modify, delete, or alter any data in the table.
 - Only respond with valid SQL queries.
+- Quote column names using double quotes (e.g., "column_name").
 - If the user asks for data visualization, respond with an SQL query to extract the necessary data for that purpose.
 - If a query involves maps, interpret this as a request for geographic map-related data and adjust your SQL query accordingly.
 - Limit your responses to information that can be derived from the database.
+- Explanations, if necessary, should be given in an SQL query of the following form, where {explanation} is the explanation: SELECT * FROM (VALUES ('{explanation}'));
 
 Do not output any natural language outside of the SQL syntax under any circumstances."
 """
